projects/ecommerce-backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# JWT Authentication
djangorestframework-simplejwt==5.3.2

# API Documentation
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
pillow==10.1.0
requests==2.31.0
